+++ package.json
@@ -15,7 +15,7 @@
     "@react-navigation/native": "^6.1.18",
     "@react-navigation/stack": "^6.4.1",
     "@reduxjs/toolkit": "^2.2.7",
-    "axios": "1.6.0",
+    "axios": "1.7.7",
     "react": "18.3.1",
     "react-native": "0.75.4",
     "react-native-gesture-handler": "^2.20.0",
