+++ src/screens/artist-detail/ArtistDetail.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import Loading from '../../components/loading/Loading';
 import styles from './ArtistDetailStyle'
 import {Image, Text, View, ScrollView} from "react-native";
-import {Constants} from "../../appconstants/AppConstants";
+import {Constants} from "../../constant/AppConstants";
 import {useGetAristDetailQuery} from "../../redux/query/RTKQuery.ts";
 import {RouteProp, useRoute} from "@react-navigation/native";
 // Define the type for the route parameters
